+++ CHANGELOG.md
@@ -5,6 +5,9 @@ CRUDlex Changelog
 Released: Upcoming
 - Attention: Removed the prefix "CRUD" from all classes as they live in their own namespace anyway
 - Attention: The data types "int" and "bool" got renamed to "integer" and "boolean"
+- Attention, API changes:
+    - EntityDefinition::getInitialSortAscending() -> EntityDefinition::isInitialSortAscending()
+    - ServiceProvider::getMangeI18N() -> ServiceProvider::isManagingI18n()
 - Replaced the markdown manual and the APIGen documentation with an unified Sphinx version
 - Added RTL support in the i18n system
 - Added file handling events